+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ---
 
 - Fixed identity switch on replies and forwards
+- Fixed draft save and send race condition
 
 
 **v0.1.38 - 03-22-2026**